Logging: Get context from previous Obsuild's message
If the context has changed since the previous message, print a header to
prefix the logs. This will help the user to differentiate which logs
belongs to which stage/pipeline.

Json.unmarshal does not error out by default if the json's content
mismatch the object's structure. Enforce strict mode to decode the
Log messages. To ensure legacy compatibility keep using the
unmarshalling method to get the Osbuild's result.