osbuild2/rpm: do not import GPG keys if repo gpgcheck is disabled
DNF imports the GPG key specified in the repository only if it can not
verify a package using already imported keys. If the `gpgcheck` is
disabled for the repository, the installed packages are not verified
and therefore the GPG key is not imported. Mimic this DNF behavior
and import the repository GPG key only if `gpgcheck` is enabled for
it.

Add unit test for `NewRPMStageOptions()` function.

Signed-off-by: Tomas Hozza <thozza@redhat.com>