worker/server: support explicitly interrupting jobs
In case a worker knows it is about to be shut down and has to discard
the current job, allow it to notify composer so the job can be immediately
requeued, rather than wait for the heartbeat to expire.

Consider this part of normal operations, and allow more retries when a
job was explicitly interrupetd, as opposed to a missing heartbeat.